requirements.txt
aiogram==3.4.0
fastapi==0.108.0
uvicorn==0.23.1
gunicorn==21.2.0
python-dotenv==1.0.0
aiosqlite==0.18.0
openpyxl==3.1.2
